style: move disconnect strategy to its own file
