+++ nomad/structs/structs_test.go
@@ -4,7 +4,6 @@
 package structs
 
 import (
-	"errors"
 	"fmt"
 	"net"
 	"os"
@@ -469,41 +468,6 @@ func TestJob_ValidateNullChar(t *testing.T) {
 	job.TaskGroups[0].Name = "so_much_better"
 	job.TaskGroups[0].Tasks[0].Name = "ive_had_it_with_these_\000_chars_in_these_names"
 	assert.Error(job.Validate(), "null character in task name should not validate")
-}
-
-func TestJob_Validate_DisconnectRescheduleLost(t *testing.T) {
-	ci.Parallel(t)
-
-	// Craft our speciality jobspec to test this particular use-case.
-	testDisconnectRescheduleLostJob := &Job{
-		ID:     "gh19644",
-		Name:   "gh19644",
-		Region: "global",
-		Type:   JobTypeSystem,
-		TaskGroups: []*TaskGroup{
-			{
-				Name: "cache",
-				Disconnect: &DisconnectStrategy{
-					LostAfter: 1 * time.Hour,
-					Replace:   pointer.Of(false),
-				},
-				Tasks: []*Task{
-					{
-						Name:   "redis",
-						Driver: "docker",
-						Config: map[string]interface{}{
-							"image": "redis:7",
-						},
-						LogConfig: DefaultLogConfig(),
-					},
-				},
-			},
-		},
-	}
-
-	testDisconnectRescheduleLostJob.Canonicalize()
-
-	must.NoError(t, testDisconnectRescheduleLostJob.Validate())
 }
 
 func TestJob_Warnings(t *testing.T) {
@@ -5823,500 +5787,6 @@ func TestAllocation_NextDelay(t *testing.T) {
 
 }
 
-// Test using stop_after_client_disconnect, remove after its deprecated  in favor
-// of Disconnect.StopAfterOnClient introduced in 1.8.0.
-func TestAllocation_WaitClientStop(t *testing.T) {
-	ci.Parallel(t)
-	type testCase struct {
-		desc                   string
-		stop                   time.Duration
-		status                 string
-		expectedShould         bool
-		expectedRescheduleTime time.Time
-	}
-	now := time.Now().UTC()
-	testCases := []testCase{
-		{
-			desc:           "running",
-			stop:           2 * time.Second,
-			status:         AllocClientStatusRunning,
-			expectedShould: true,
-		},
-		{
-			desc:           "no stop_after_client_disconnect",
-			status:         AllocClientStatusLost,
-			expectedShould: false,
-		},
-		{
-			desc:                   "stop",
-			status:                 AllocClientStatusLost,
-			stop:                   2 * time.Second,
-			expectedShould:         true,
-			expectedRescheduleTime: now.Add((2 + 5) * time.Second),
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.desc, func(t *testing.T) {
-			j := testJob()
-			a := &Allocation{
-				ClientStatus: tc.status,
-				Job:          j,
-				TaskStates:   map[string]*TaskState{},
-			}
-
-			if tc.status == AllocClientStatusLost {
-				a.AppendState(AllocStateFieldClientStatus, AllocClientStatusLost)
-			}
-
-			j.TaskGroups[0].StopAfterClientDisconnect = &tc.stop
-			a.TaskGroup = j.TaskGroups[0].Name
-
-			require.Equal(t, tc.expectedShould, a.ShouldClientStop())
-
-			if !tc.expectedShould || tc.status != AllocClientStatusLost {
-				return
-			}
-
-			// the reschedTime is close to the expectedRescheduleTime
-			reschedTime := a.WaitClientStop()
-			e := reschedTime.Unix() - tc.expectedRescheduleTime.Unix()
-			require.Less(t, e, int64(2))
-		})
-	}
-}
-
-func TestAllocation_WaitClientStop_Disconnect(t *testing.T) {
-	ci.Parallel(t)
-	type testCase struct {
-		desc                   string
-		stop                   time.Duration
-		status                 string
-		expectedShould         bool
-		expectedRescheduleTime time.Time
-	}
-	now := time.Now().UTC()
-	testCases := []testCase{
-		{
-			desc:           "running",
-			stop:           2 * time.Second,
-			status:         AllocClientStatusRunning,
-			expectedShould: true,
-		},
-		{
-			desc:           "no stop_after_client_disconnect",
-			status:         AllocClientStatusLost,
-			expectedShould: false,
-		},
-		{
-			desc:                   "stop",
-			status:                 AllocClientStatusLost,
-			stop:                   2 * time.Second,
-			expectedShould:         true,
-			expectedRescheduleTime: now.Add((2 + 5) * time.Second),
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.desc, func(t *testing.T) {
-			j := testJob()
-			a := &Allocation{
-				ClientStatus: tc.status,
-				Job:          j,
-				TaskStates:   map[string]*TaskState{},
-			}
-
-			if tc.status == AllocClientStatusLost {
-				a.AppendState(AllocStateFieldClientStatus, AllocClientStatusLost)
-			}
-
-			j.TaskGroups[0].Disconnect = &DisconnectStrategy{
-				StopAfterOnClient: &tc.stop,
-			}
-
-			a.TaskGroup = j.TaskGroups[0].Name
-
-			require.Equal(t, tc.expectedShould, a.ShouldClientStop())
-
-			if !tc.expectedShould || tc.status != AllocClientStatusLost {
-				return
-			}
-
-			// the reschedTime is close to the expectedRescheduleTime
-			reschedTime := a.WaitClientStop()
-			e := reschedTime.Unix() - tc.expectedRescheduleTime.Unix()
-			require.Less(t, e, int64(2))
-		})
-	}
-}
-
-func TestAllocation_Timeout_Disconnect(t *testing.T) {
-	type testCase struct {
-		desc          string
-		maxDisconnect time.Duration
-	}
-
-	testCases := []testCase{
-		{
-			desc:          "has lost_after",
-			maxDisconnect: 30 * time.Second,
-		},
-		{
-			desc:          "zero lost_after",
-			maxDisconnect: 0 * time.Second,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.desc, func(t *testing.T) {
-			j := testJob()
-			a := &Allocation{
-				Job: j,
-			}
-
-			j.TaskGroups[0].Disconnect = &DisconnectStrategy{
-				LostAfter: tc.maxDisconnect,
-			}
-
-			a.TaskGroup = j.TaskGroups[0].Name
-
-			now := time.Now()
-
-			reschedTime := a.DisconnectTimeout(now)
-
-			if tc.maxDisconnect == 0 {
-				require.Equal(t, now, reschedTime, "expected to be now")
-			} else {
-				difference := reschedTime.Sub(now)
-				require.Equal(t, tc.maxDisconnect, difference, "expected durations to be equal")
-			}
-
-		})
-	}
-}
-
-// Test using max_client_disconnect, remove after its deprecated  in favor
-// of Disconnect.LostAfter introduced in 1.8.0.
-func TestAllocation_DisconnectTimeout(t *testing.T) {
-	type testCase struct {
-		desc          string
-		maxDisconnect *time.Duration
-	}
-
-	testCases := []testCase{
-		{
-			desc:          "no max_client_disconnect",
-			maxDisconnect: nil,
-		},
-		{
-			desc:          "has max_client_disconnect",
-			maxDisconnect: pointer.Of(30 * time.Second),
-		},
-		{
-			desc:          "zero max_client_disconnect",
-			maxDisconnect: pointer.Of(0 * time.Second),
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.desc, func(t *testing.T) {
-			j := testJob()
-			a := &Allocation{
-				Job: j,
-			}
-
-			j.TaskGroups[0].MaxClientDisconnect = tc.maxDisconnect
-			a.TaskGroup = j.TaskGroups[0].Name
-
-			now := time.Now()
-
-			reschedTime := a.DisconnectTimeout(now)
-
-			if tc.maxDisconnect == nil {
-				require.Equal(t, now, reschedTime, "expected to be now")
-			} else {
-				difference := reschedTime.Sub(now)
-				require.Equal(t, *tc.maxDisconnect, difference, "expected durations to be equal")
-			}
-		})
-	}
-}
-
-// Test using max_client_disconnect, remove after its deprecated  in favor
-// of Disconnect.LostAfter introduced in 1.8.0.
-func TestAllocation_Expired(t *testing.T) {
-	type testCase struct {
-		name             string
-		maxDisconnect    string
-		ellapsed         int
-		expected         bool
-		nilJob           bool
-		badTaskGroup     bool
-		mixedUTC         bool
-		noReconnectEvent bool
-		status           string
-	}
-
-	testCases := []testCase{
-		{
-			name:          "has-expired",
-			maxDisconnect: "5s",
-			ellapsed:      10,
-			expected:      true,
-		},
-		{
-			name:          "has-not-expired",
-			maxDisconnect: "5s",
-			ellapsed:      3,
-			expected:      false,
-		},
-		{
-			name:          "are-equal",
-			maxDisconnect: "5s",
-			ellapsed:      5,
-			expected:      true,
-		},
-		{
-			name:          "nil-job",
-			maxDisconnect: "5s",
-			ellapsed:      10,
-			expected:      false,
-			nilJob:        true,
-		},
-		{
-			name:          "wrong-status",
-			maxDisconnect: "5s",
-			ellapsed:      10,
-			expected:      false,
-			status:        AllocClientStatusRunning,
-		},
-		{
-			name:          "bad-task-group",
-			maxDisconnect: "",
-			badTaskGroup:  true,
-			ellapsed:      10,
-			expected:      false,
-		},
-		{
-			name:          "no-max-disconnect",
-			maxDisconnect: "",
-			ellapsed:      10,
-			expected:      false,
-		},
-		{
-			name:          "mixed-utc-has-expired",
-			maxDisconnect: "5s",
-			ellapsed:      10,
-			mixedUTC:      true,
-			expected:      true,
-		},
-		{
-			name:          "mixed-utc-has-not-expired",
-			maxDisconnect: "5s",
-			ellapsed:      3,
-			mixedUTC:      true,
-			expected:      false,
-		},
-		{
-			name:             "no-reconnect-event",
-			maxDisconnect:    "5s",
-			ellapsed:         2,
-			expected:         false,
-			noReconnectEvent: true,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			alloc := MockAlloc()
-			var err error
-			var maxDisconnect time.Duration
-
-			if tc.maxDisconnect != "" {
-				maxDisconnect, err = time.ParseDuration(tc.maxDisconnect)
-				require.NoError(t, err)
-				alloc.Job.TaskGroups[0].MaxClientDisconnect = &maxDisconnect
-			}
-
-			if tc.nilJob {
-				alloc.Job = nil
-			}
-
-			if tc.badTaskGroup {
-				alloc.TaskGroup = "bad"
-			}
-
-			alloc.ClientStatus = AllocClientStatusUnknown
-			if tc.status != "" {
-				alloc.ClientStatus = tc.status
-			}
-
-			alloc.AllocStates = []*AllocState{{
-				Field: AllocStateFieldClientStatus,
-				Value: AllocClientStatusUnknown,
-				Time:  time.Now(),
-			}}
-
-			require.NoError(t, err)
-			now := time.Now().UTC()
-			if tc.mixedUTC {
-				now = time.Now()
-			}
-
-			if !tc.noReconnectEvent {
-				event := NewTaskEvent(TaskClientReconnected)
-				event.Time = now.UnixNano()
-
-				alloc.TaskStates = map[string]*TaskState{
-					"web": {
-						Events: []*TaskEvent{event},
-					},
-				}
-			}
-
-			ellapsedDuration := time.Duration(tc.ellapsed) * time.Second
-			now = now.Add(ellapsedDuration)
-
-			require.Equal(t, tc.expected, alloc.Expired(now))
-		})
-	}
-}
-
-func TestAllocation_Expired_Disconnected(t *testing.T) {
-	type testCase struct {
-		name             string
-		maxDisconnect    string
-		ellapsed         int
-		expected         bool
-		nilJob           bool
-		badTaskGroup     bool
-		mixedUTC         bool
-		noReconnectEvent bool
-		status           string
-	}
-
-	testCases := []testCase{
-		{
-			name:          "has-expired",
-			maxDisconnect: "5s",
-			ellapsed:      10,
-			expected:      true,
-		},
-		{
-			name:          "has-not-expired",
-			maxDisconnect: "5s",
-			ellapsed:      3,
-			expected:      false,
-		},
-		{
-			name:          "are-equal",
-			maxDisconnect: "5s",
-			ellapsed:      5,
-			expected:      true,
-		},
-		{
-			name:          "nil-job",
-			maxDisconnect: "5s",
-			ellapsed:      10,
-			expected:      false,
-			nilJob:        true,
-		},
-		{
-			name:          "wrong-status",
-			maxDisconnect: "5s",
-			ellapsed:      10,
-			expected:      false,
-			status:        AllocClientStatusRunning,
-		},
-		{
-			name:          "bad-task-group",
-			maxDisconnect: "",
-			badTaskGroup:  true,
-			ellapsed:      10,
-			expected:      false,
-		},
-		{
-			name:          "no-max-disconnect",
-			maxDisconnect: "",
-			ellapsed:      10,
-			expected:      false,
-		},
-		{
-			name:          "mixed-utc-has-expired",
-			maxDisconnect: "5s",
-			ellapsed:      10,
-			mixedUTC:      true,
-			expected:      true,
-		},
-		{
-			name:          "mixed-utc-has-not-expired",
-			maxDisconnect: "5s",
-			ellapsed:      3,
-			mixedUTC:      true,
-			expected:      false,
-		},
-		{
-			name:             "no-reconnect-event",
-			maxDisconnect:    "5s",
-			ellapsed:         2,
-			expected:         false,
-			noReconnectEvent: true,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			alloc := MockAlloc()
-			var err error
-			var maxDisconnect time.Duration
-
-			if tc.maxDisconnect != "" {
-				maxDisconnect, err = time.ParseDuration(tc.maxDisconnect)
-				require.NoError(t, err)
-				alloc.Job.TaskGroups[0].Disconnect = &DisconnectStrategy{
-					LostAfter: maxDisconnect,
-				}
-			}
-
-			if tc.nilJob {
-				alloc.Job = nil
-			}
-
-			if tc.badTaskGroup {
-				alloc.TaskGroup = "bad"
-			}
-
-			alloc.ClientStatus = AllocClientStatusUnknown
-			if tc.status != "" {
-				alloc.ClientStatus = tc.status
-			}
-
-			alloc.AllocStates = []*AllocState{{
-				Field: AllocStateFieldClientStatus,
-				Value: AllocClientStatusUnknown,
-				Time:  time.Now(),
-			}}
-
-			require.NoError(t, err)
-			now := time.Now().UTC()
-			if tc.mixedUTC {
-				now = time.Now()
-			}
-
-			if !tc.noReconnectEvent {
-				event := NewTaskEvent(TaskClientReconnected)
-				event.Time = now.UnixNano()
-
-				alloc.TaskStates = map[string]*TaskState{
-					"web": {
-						Events: []*TaskEvent{event},
-					},
-				}
-			}
-
-			ellapsedDuration := time.Duration(tc.ellapsed) * time.Second
-			now = now.Add(ellapsedDuration)
-
-			require.Equal(t, tc.expected, alloc.Expired(now))
-		})
-	}
-}
-
 func TestAllocation_NeedsToReconnect(t *testing.T) {
 	ci.Parallel(t)
 
@@ -6597,119 +6067,6 @@ func TestParameterizedJobConfig_Validate_NonBatch(t *testing.T) {
 	if err := job.Validate(); err == nil || !strings.Contains(err.Error(), "only be used with") {
 		t.Fatalf("Expected bad scheduler tpye: %v", err)
 	}
-}
-
-// Test using stop_after_client_disconnect, remove after its deprecated  in favor
-// of Disconnect.StopAfterOnClient introduced in 1.8.0.
-func TestJobConfig_Validate_StopAferClientDisconnect(t *testing.T) {
-	ci.Parallel(t)
-	// Setup a system Job with stop_after_client_disconnect set, which is invalid
-	job := testJob()
-	job.Type = JobTypeSystem
-	stop := 1 * time.Minute
-	job.TaskGroups[0].StopAfterClientDisconnect = &stop
-
-	err := job.Validate()
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "stop_after_client_disconnect can only be set in batch and service jobs")
-
-	// Modify the job to a batch job with an invalid stop_after_client_disconnect value
-	job.Type = JobTypeBatch
-	invalid := -1 * time.Minute
-	job.TaskGroups[0].StopAfterClientDisconnect = &invalid
-
-	err = job.Validate()
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "stop_after_client_disconnect must be a positive value")
-
-	// Modify the job to a batch job with a valid stop_after_client_disconnect value
-	job.Type = JobTypeBatch
-	job.TaskGroups[0].StopAfterClientDisconnect = &stop
-	err = job.Validate()
-	require.NoError(t, err)
-}
-
-func TestJobConfig_Validate_StopAferClient_Disconnect(t *testing.T) {
-	ci.Parallel(t)
-	// Setup a system Job with Disconnect.StopAfterOnClient set, which is invalid
-	job := testJob()
-	job.Type = JobTypeSystem
-	stop := 1 * time.Minute
-	job.TaskGroups[0].Disconnect = &DisconnectStrategy{
-		StopAfterOnClient: &stop,
-	}
-
-	err := job.Validate()
-	require.Error(t, err)
-	require.Contains(t, err.Error(), errStopAfterNonService.Error())
-
-	// Modify the job to a batch job with an invalid Disconnect.StopAfterOnClient value
-	job.Type = JobTypeBatch
-	invalid := -1 * time.Minute
-	job.TaskGroups[0].Disconnect = &DisconnectStrategy{
-		StopAfterOnClient: &invalid,
-	}
-
-	err = job.Validate()
-	require.Error(t, err)
-	require.Contains(t, err.Error(), errNegativeStopAfter.Error())
-
-	// Modify the job to a batch job with a valid Disconnect.StopAfterOnClient value
-	job.Type = JobTypeBatch
-	job.TaskGroups[0].Disconnect = &DisconnectStrategy{
-		StopAfterOnClient: &stop,
-	}
-	err = job.Validate()
-	require.NoError(t, err)
-}
-
-// Test using max_client_disconnect, remove after its deprecated  in favor
-// of Disconnect.LostAfter introduced in 1.8.0.
-func TestJobConfig_Validate_MaxClientDisconnect(t *testing.T) {
-	// Set up a job with an invalid max_client_disconnect value
-	job := testJob()
-	timeout := -1 * time.Minute
-	job.TaskGroups[0].MaxClientDisconnect = &timeout
-	job.TaskGroups[0].StopAfterClientDisconnect = &timeout
-
-	err := job.Validate()
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "max_client_disconnect cannot be negative")
-	require.Contains(t, err.Error(), "Task group cannot be configured with both max_client_disconnect and stop_after_client_disconnect")
-
-	// Modify the job with a valid max_client_disconnect value
-	timeout = 1 * time.Minute
-	job.TaskGroups[0].MaxClientDisconnect = &timeout
-	job.TaskGroups[0].StopAfterClientDisconnect = nil
-	err = job.Validate()
-	require.NoError(t, err)
-}
-
-func TestJobConfig_Validate_LostAfter_Disconnect(t *testing.T) {
-	// Set up a job with an invalid Disconnect.LostAfter value
-	job := testJob()
-	timeout := -1 * time.Minute
-	job.TaskGroups[0].Disconnect = &DisconnectStrategy{
-		LostAfter:         timeout,
-		StopAfterOnClient: &timeout,
-	}
-
-	err := job.Validate()
-	require.Error(t, err)
-	err = errors.Unwrap(err)
-
-	require.Contains(t, err.Error(), errNegativeLostAfter.Error())
-	require.Contains(t, err.Error(), errNegativeStopAfter.Error())
-	require.Contains(t, err.Error(), errStopAndLost.Error())
-
-	// Modify the job with a valid Disconnect.LostAfter value
-	timeout = 1 * time.Minute
-	job.TaskGroups[0].Disconnect = &DisconnectStrategy{
-		LostAfter:         timeout,
-		StopAfterOnClient: nil,
-	}
-	err = job.Validate()
-	require.NoError(t, err)
 }
 
 func TestParameterizedJobConfig_Canonicalize(t *testing.T) {
@@ -8814,83 +8171,6 @@ func TestNewIdentityClaims(t *testing.T) {
 			must.Eq(t, alloc.ID, got.AllocationID)
 			must.Eq(t, jwt.NewNumericDate(now), got.IssuedAt)
 			must.Eq(t, jwt.NewNumericDate(now), got.NotBefore)
-		})
-	}
-}
-
-func TestDisconnectStategy_Validate(t *testing.T) {
-	ci.Parallel(t)
-
-	cases := []struct {
-		name     string
-		strategy *DisconnectStrategy
-		jobType  string
-		err      error
-	}{
-		{
-			name: "negative-stop-after",
-			strategy: &DisconnectStrategy{
-				StopAfterOnClient: pointer.Of(-1 * time.Second),
-			},
-			jobType: JobTypeService,
-			err:     errNegativeStopAfter,
-		},
-		{
-			name: "stop-after-on-system",
-			strategy: &DisconnectStrategy{
-				StopAfterOnClient: pointer.Of(1 * time.Second),
-			},
-			jobType: JobTypeSystem,
-			err:     errStopAfterNonService,
-		},
-		{
-			name: "negative-lost-after",
-			strategy: &DisconnectStrategy{
-				LostAfter: -1 * time.Second,
-			},
-			jobType: JobTypeService,
-			err:     errNegativeLostAfter,
-		},
-		{
-			name: "lost-after-and-stop-after-enabled",
-			strategy: &DisconnectStrategy{
-				LostAfter:         1 * time.Second,
-				StopAfterOnClient: pointer.Of(1 * time.Second),
-			},
-			jobType: JobTypeService,
-			err:     errStopAndLost,
-		},
-		{
-			name: "invalid-reconcile",
-			strategy: &DisconnectStrategy{
-				LostAfter: 1 * time.Second,
-				Reconcile: "invalid",
-			},
-			jobType: JobTypeService,
-			err:     errInvalidReconcile,
-		},
-		{
-			name: "valid-configuration",
-			strategy: &DisconnectStrategy{
-				LostAfter:         1 * time.Second,
-				Reconcile:         ReconcileOptionKeepOriginal,
-				Replace:           pointer.Of(true),
-				StopAfterOnClient: nil,
-			},
-			jobType: JobTypeService,
-			err:     nil,
-		},
-	}
-
-	job := testJob()
-
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			job.Type = c.jobType
-			err := c.strategy.Validate(job)
-			if !errors.Is(err, c.err) {
-				t.Errorf("expected error %v, got %v", c.err, err)
-			}
 		})
 	}
 }
